included logo and version in both HNQIS and EDS
