+++ app/src/eds/java/org/eyeseetea/malariacare/utils/Utils.java
@@ -43,16 +43,6 @@ import org.eyeseetea.malariacare.network.PullClient;
 public class Utils extends AUtils {
 
     @Override
-    public void showAlert(int titleId, CharSequence text, Context context){
-        final AlertDialog dialog = new AlertDialog.Builder(context)
-                .setTitle(context.getString(titleId))
-                .setMessage(text)
-                .setNeutralButton(android.R.string.ok, null).create();
-        dialog.show();
-        ((TextView)dialog.findViewById(android.R.id.message)).setMovementMethod(LinkMovementMethod.getInstance());
-    }
-
-    @Override
     public void createNewSurvey(OrgUnit orgUnit, TabGroup tabGroup) {
         ComboOrgUnitTabGroup comboOrgUnitTabGroup=new ComboOrgUnitTabGroup(orgUnit,tabGroup);
         LoadLastEvent loadLastEvent= new LoadLastEvent();
